requirements.txt
fastapi==0.135.3
httpx==0.28.1
itsdangerous==2.2.0
Jinja2==3.1.6
python-dotenv==1.2.2
Requests==2.33.1
TikTokApi==7.3.3
starlette<1.0.0
